+++ waymore/headers/Camera.h
@@ -13,7 +13,7 @@ typedef void Camera; // Intermediate for C compatibility
 
 Camera * initializeCamera(const int pixelWidth, const int pixelHeight, const int slices);
 void startCamera();
-void getCameraLineDistances(int * distanceBuffer);
+void getCameraLineDistances(double * distanceBuffer);
 void getCameraLineConfidences(double * confidenceBuffer);
 void uninitializeCamera();
